Change test for k_best to be class numeric

+++ tests/testthat/test_multi_clustering.R
@@ -66,6 +66,6 @@ test_that("making sure multiClust object works properly", {
   expect_equal(length(f_clust[["wss"]]), 4)
   expect_is(f_clust[["clust_gap"]], "clusGap")
   expect_equal(length(f_clust[["clust_gap"]]), 4)
-  expect_is(f_clust[["k_best"]], "integer")
+  expect_is(f_clust[["k_best"]], "numeric")
   expect_equal(length(f_clust[["k_best"]]), 1)
 })
